+++ src/LibPARI.jl
@@ -18,6 +18,7 @@ include("lifecycle.jl")
 include("gen.jl")
 include("errors.jl")
 include("conversions.jl")
+include("numeric.jl")
 include("bindings.jl")
 
 """